feat: `useTimer` 커스텀훅 추가
